requirements.txt
flask>=3.0
flask-cors>=4.0
anthropic>=0.40
python-dotenv>=1.0
gunicorn>=21.0
supabase>=2.0
